backend/requirements.txt
# Core FastAPI
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.23
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.1

# Basic ML and Image Processing
pillow>=10.3.0
numpy>=1.26.0
scikit-learn>=1.3.2

# Google Gemini
google-generativeai>=0.3.1

# Utilities
aiofiles>=23.2.1
python-dateutil>=2.8.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.2

# Development
black>=23.12.0
flake8>=6.1.0

# Production ML Features
onnxruntime>=1.16.3
opencv-python-headless>=4.8.1.78
lime>=0.2.0.1
